+++ RELEASE-NOTES.md
@@ -18,6 +18,7 @@ as release notes for end user on a Recalbox upgrade.
 - Add game Pong Recalbox Edition by Studio 40A on gameboy
 - Add HDMI priority option in recalbox rgb dual to switch between Scart and HDMI
 - Add guncon2 support for CRT tvs
+- Fix resolution selection in frontend for Raspberry Pi 4/400/3/Zero2 (KMS)
 
 ### Improvements
 - Bump libretro pcsx-rearmed